feat: update release scripts
