+++ .gitignore
@@ -7,3 +7,7 @@ api_samples/
 .envrc
 mutants.out/
 mutants.out.old/
+
+/templates/release_scripts/enum.txt
+/templates/release_scripts/name.txt
+/templates/release_scripts/code.txt
