+++ src/entity/mod.rs
@@ -96,6 +96,7 @@ pub mod recording;
 pub mod relations;
 pub mod release;
 pub mod release_group;
+pub mod release_scripts;
 pub mod search;
 pub mod series;
 pub mod tag;
