+++ templates/release_scripts/file_template.txt
@@ -1,0 +1,49 @@
+/// The script used to write the release's track list. The possible values are taken from the
+/// [ISO 15924](https://en.wikipedia.org/wiki/ISO_15924) standard.
+///
+/// The values for this enum have been generated with the following command:
+///
+/// ```bash
+/// $ curl -s https://musicbrainz.org/statistics/languages-scripts | \
+///     grep -Eo '<td>[^<]*</td><td class="t"><a href="https://musicbrainz.org/search\?query=script%3A%22[^"]*%22' | \
+///     sort | \
+///     sed 's,<td>\([^<]*\)</td><td class="t"><a href="https://musicbrainz.org/search?query=script%3A%22\([^"]*\)%22,\/\/\/ \1\n\2\,,'
+/// ```
+#[derive(Debug, serde::Serialize, serde::Deserialize, PartialEq, Eq, Clone)]
+pub enum ReleaseScript {
+
+}
+
+impl ReleaseScript {
+    /// Get the human-readable name used by MusicBrainz.
+    ///
+    /// The values for this enum have been generated with the following command:
+    ///
+    /// ```bash
+    /// $ curl -s https://musicbrainz.org/statistics/languages-scripts | \
+    ///     grep -Eo '<td>[^<]*</td><td class="t"><a href="https://musicbrainz.org/search\?query=script%3A%22[^"]*%22' | \
+    ///     sort | \
+    ///     sed 's,<td>\([^<]*\)</td><td class="t"><a href="https://musicbrainz.org/search?query=script%3A%22\([^"]*\)%22,            Self::\2 => "\1"\,,'
+    /// ```
+    pub fn name(&self) -> &'static str {
+        match &self {
+
+        }
+    }
+
+    /// Get the [ISO 15924](https://en.wikipedia.org/wiki/ISO_15924) code as [`str`].
+    ///
+    /// The values for this enum have been generated with the following command:
+    ///
+    /// ```bash
+    /// $ curl -s https://musicbrainz.org/statistics/languages-scripts | \
+    ///     grep -Eo '<td>[^<]*</td><td class="t"><a href="https://musicbrainz.org/search\?query=script%3A%22[^"]*%22' | \
+    ///     sort | \
+    ///     sed 's,<td>\([^<]*\)</td><td class="t"><a href="https://musicbrainz.org/search?query=script%3A%22\([^"]*\)%22,            Self::\2 => "\2"\,,'
+    /// ```
+    pub fn code(&self) -> &'static str {
+        match &self {
+
+        }
+    }
+}
